somewhat responsive design
